+++ RecipeBook/app/src/main/java/uv/recipebook/data/Ingredient.kt
@@ -4,5 +4,5 @@ import java.io.Serializable
 
 data class Ingredient (
     var ingredient : String,
-    var amountInGr : Int
+    var amountInGr : Double
 ) : Serializable
